requirements.txt
streamlit
numpy
opencv-python
scikit-image
Pillow
joblib
gdown
